requirements.txt
ail-typo-squatting==2.7.4
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.2
cssselect==1.2.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
dotenv==0.9.9
executing==2.2.0
fastapi==0.115.11
filelock==3.17.0
firebase-admin==6.7.0
fuzzywuzzy==0.18.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
ImageHash==4.3.2
incremental==24.7.2
inflect==5.6.2
ipykernel==6.29.5
# ipython==9.0.2
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
joblib==1.4.2
json2html==1.3.0
jupyter_client==8.6.3
jupyter_core==5.7.2
Levenshtein==0.27.1
lightgbm==4.6.0
lxml==5.3.1
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.5
pytz==2025.1
PyWavelets==1.8.0
# pywin32==309
PyYAML==6.0.2
pyzmq==26.2.1
queuelib==1.7.0
RapidFuzz==3.12.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
retrie==0.2.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.9.1
six==1.17.0
sniffio==1.3.1
socks==0
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.1
threadpoolctl==3.5.0
tldextract==5.1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
w3lib==2.3.1
wcwidth==0.2.13
zope.interface==7.2
